Rename io module to config
